# endpoints: remove obsolete ServiceSelectorCache